+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/traces/generate_trace.ts
