+++ x-pack/test_serverless/api_integration/test_suites/observability/index.ts
@@ -12,7 +12,6 @@ export default function ({ loadTestFile }: FtrProviderContext) {
     this.tags(['esGate']);
 
     loadTestFile(require.resolve('./apm_api_integration/feature_flags.ts'));
-    loadTestFile(require.resolve('./apm_api_integration/traces/critical_path'));
     loadTestFile(require.resolve('./cases'));
     loadTestFile(require.resolve('./synthetics'));
     loadTestFile(require.resolve('./dataset_quality_api_integration'));
